Move constants and errors to the lib.rs module
